--- src/Application/Platforms/Commands/AddAccount/AddAccount.cs
@@ -1,7 +1,0 @@
-﻿namespace GitNode.Application.Platforms.Commands.AddAccount
-{
-    public class AddAccount
-    {
-        public string Code { get; set; }
-    }
-}
