+++ src/Application/Platforms/Commands/CreateRepository/CreateRepositoryCommand.cs
@@ -1,22 +1,13 @@
-﻿using GitNode.Application.Common.Dtos;
-using GitNode.Application.Common.Interfaces;
-using GitNode.Application.Common.Mappings;
-using MediatR;
+﻿using System.Collections.Generic;
 
 namespace GitNode.Application.Platforms.Commands.CreateRepository
 {
-    public class CreateRepositoryCommand : CreateRepository,
-                                           IRequest<PlatformRepositoryDto>,
-                                           IMapFrom<CreateRepository>,
-                                           IPlatform
+    public class CreateRepository
     {
-        public CreateRepositoryCommand() { }
-
-        public CreateRepositoryCommand(string platform)
-        {
-            Platform = platform;
-        }
-
-        public string Platform { get; set; }
+        public string Username { get; set; }
+        public string RepoName { get; set; }
+        public string Description { get; set; }
+        public bool Private { get; set; }
+        public IEnumerable<string> Technologies { get; set; }
     }
 }
