+++ src/Application/Platforms/Commands/CreateRepository/CreateRepositoryExtendedCommand.cs
@@ -9,13 +9,28 @@ using GitNode.Application.Common.Data.ExternalAPIs;
 using GitNode.Application.Common.Dtos;
 using GitNode.Application.Common.Exceptions;
 using GitNode.Application.Common.Interfaces;
-using GitNode.Application.Platforms.Commands.CreateRepository;
+using GitNode.Application.Common.Mappings;
 using GitNode.Domain.Entities;
 using MediatR;
 
-namespace GitNode.Application.Platforms.Handlers.CommandHandlers
+namespace GitNode.Application.Platforms.Commands.CreateRepository
 {
-    class CreateRepositoryCommandHandler : PlatformHandlerBase,
+    public class CreateRepositoryCommand : CreateRepository,
+                                           IRequest<PlatformRepositoryDto>,
+                                           IMapFrom<CreateRepository>,
+                                           IPlatform
+    {
+        public CreateRepositoryCommand() { }
+
+        public CreateRepositoryCommand(string platform)
+        {
+            Platform = platform;
+        }
+
+        public string Platform { get; set; }
+    }
+    
+    public class CreateRepositoryCommandHandler : PlatformHandlerBase,
         IRequestHandler<CreateRepositoryCommand, PlatformRepositoryDto>
     {
         private readonly ICurrentUserService _currentUser;
